清理代码

+++ src/Magicodes.ExporterAndImporter.Tests/PdfExporter_Tests.cs
@@ -1,9 +1,24 @@
-﻿using GenFu;
+﻿// ======================================================================
+// 
+//           Copyright (C) 2019-2030 湖南心莱信息科技有限公司
+//           All rights reserved
+// 
+//           filename : PdfExporter_Tests.cs
+//           description :
+// 
+//           created by 雪雁 at  2019-09-26 14:59
+//           文档官网：https://docs.xin-lai.com
+//           公众号教程：麦扣聊技术
+//           QQ群：85318032（编程交流）
+//           Blog：http://www.cnblogs.com/codelove/
+// 
+// ======================================================================
+
+using System.IO;
+using System.Threading.Tasks;
 using Magicodes.ExporterAndImporter.Pdf;
 using Magicodes.ExporterAndImporter.Tests.Models;
 using Shouldly;
-using System.IO;
-using System.Threading.Tasks;
 using Xunit;
 
 namespace Magicodes.ExporterAndImporter.Tests
@@ -15,12 +30,9 @@ namespace Magicodes.ExporterAndImporter.Tests
         {
             var exporter = new PdfExporter();
             var filePath = Path.Combine(Directory.GetCurrentDirectory(), nameof(ExportPdf_Test) + ".pdf");
-            if (File.Exists(filePath))
-            {
-                File.Delete(filePath);
-            }
+            if (File.Exists(filePath)) File.Delete(filePath);
             //此处使用默认模板导出
-            var result = await exporter.ExportByTemplate(filePath, A.ListOf<ExportTestData>());
+            var result = await exporter.ExportByTemplate(filePath, GenFu.GenFu.ListOf<ExportTestData>());
             result.ShouldNotBeNull();
             File.Exists(filePath).ShouldBeTrue();
         }
@@ -33,13 +45,10 @@ namespace Magicodes.ExporterAndImporter.Tests
             var tpl = File.ReadAllText(tplPath);
             var exporter = new PdfExporter();
             var filePath = Path.Combine(Directory.GetCurrentDirectory(), nameof(ExportPdfByTemplate_Test) + ".pdf");
-            if (File.Exists(filePath))
-            {
-                File.Delete(filePath);
-            }
+            if (File.Exists(filePath)) File.Delete(filePath);
             //此处使用默认模板导出
             var result = await exporter.ExportByTemplate(filePath,
-                A.ListOf<ExportTestData>(), tpl);
+                GenFu.GenFu.ListOf<ExportTestData>(), tpl);
             result.ShouldNotBeNull();
             File.Exists(filePath).ShouldBeTrue();
         }
